separate into some test js
